+++ Backend/src/main/java/com/bodegaaurrera/perecederos_demo/DTO/DiscrepanciaDTO.java
@@ -2,10 +2,12 @@ package com.bodegaaurrera.perecederos_demo.DTO;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class DiscrepanciaDTO {
     private Long idDiscrepancia;
-    private int totalEsperado;   // 🔹 nuevo
-    private int totalRecibido;   // 🔹 nuevo
-    private int totalFaltante;
+    private BigDecimal totalEsperado;   // 🔹 nuevo
+    private BigDecimal totalRecibido;   // 🔹 nuevo
+    private BigDecimal totalFaltante;
 }
