+++ Backend/src/main/java/com/bodegaaurrera/perecederos_demo/DTO/InventarioDTO.java
@@ -3,6 +3,7 @@ package com.bodegaaurrera.perecederos_demo.DTO;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 
@@ -11,7 +12,7 @@ import java.time.LocalDate;
 public class InventarioDTO {
     private String codigoBarras;
     private String descripcion;
-    private int cantidad;
+    private BigDecimal cantidad;
     private String lote;
     private LocalDate fechaCaducidad;
 }
