+++ Backend/src/main/java/com/bodegaaurrera/perecederos_demo/DTO/InventarioDetalleDTO.java
@@ -3,6 +3,7 @@ package com.bodegaaurrera.perecederos_demo.DTO;
 import lombok.Data;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,9 +11,9 @@ public class InventarioDetalleDTO {
 
         private ProductoDTO producto;
 
-        private int inventarioIp;
-        private int pisoVenta;
-        private int bodega;
+        private BigDecimal inventarioIp;
+        private BigDecimal pisoVenta;
+        private BigDecimal bodega;
 
         private List<LoteDTO> lotes;
