+++ Backend/src/main/java/com/bodegaaurrera/perecederos_demo/DTO/LoteMovimientoDTO.java
@@ -3,10 +3,12 @@ package com.bodegaaurrera.perecederos_demo.DTO;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class LoteMovimientoDTO {
 
     private String lote;
-    private int cantidad;
+    private BigDecimal cantidad;
 }
